completely working crawler
